feat: 상품 상세조회

+++ app/src/main/java/com/lastminute/ui/model/ProductDetail.kt
@@ -3,14 +3,14 @@ package com.lastminute.ui.model
 import java.time.LocalDateTime
 
 data class ProductDetail(
-    val productId: Int,
+    val productId: Long,
     val title: String,
     val menu: String,
     val numPeoples: Int,
     val address: String,
     val lat: Double,
     val lot: Double,
-    val writerId: Int,
+    val writerId: Long,
     val writer: String,
     val time: LocalDateTime,
     val isAllPaid: Boolean,
